+++ src/reducer/data/data.js
@@ -15,48 +15,6 @@ const ActionCreator = {
   })
 };
 
-const OperationOffers = {
-  loadOffers: () => (dispatch, getState, api) => {
-    return api.get(`/hotels`).then((response) => {
-      const myOffers = response.data.reduce((result, offer) => {
-        if (!result[offer.city.name]) {
-          result[offer.city.name] = {
-            city: offer.city,
-            offers: []
-          };
-        }
-        result[offer.city.name][`offers`].push(offer);
-
-        return result;
-      }, {});
-      const cities = Object.keys(myOffers).sort();
-      cities.map((city)=>{
-        const renameKeysArrayOffers = Array.from(myOffers[city].offers, (offer)=> {
-          const newObj = Object.assign({}, offer, {
-            previewImage: offer.preview_image,
-            isFavorite: offer.is_favorite,
-            isPremium: offer.is_premium,
-            maxAdults: offer.max_adults,
-            hostId: offer.host.id,
-            hostName: offer.host.name,
-            hostIsPro: offer.host.is_pro,
-            hostAvatarUrl: offer.host.avatar_url
-          });
-          delete newObj.city;
-          delete newObj.preview_image;
-          delete newObj.is_favorite;
-          delete newObj.is_premium;
-          delete newObj.max_adults;
-          delete newObj.host;
-          return newObj;
-        });
-        myOffers[city].offers = renameKeysArrayOffers;
-      });
-      dispatch(ActionCreator.loadOffers(myOffers));
-    });
-  }
-};
-
 const reducer = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.LOAD_OFFERS:
@@ -68,6 +26,5 @@ const reducer = (state = initialState, action) => {
 export {
   reducer,
   ActionType,
-  OperationOffers,
   ActionCreator,
 };
