+++ src/reducer/data/selectors.js
@@ -1,6 +1,8 @@
-const getCities = (state) => Object.keys(state.offers).sort();
+import NameSpace from "../name-space.js";
+
+const getCities = (state) => Object.keys(state[NameSpace.DATA].offers).sort();
 const getCityOffers = (state) => {
-  return state.offers[state.city];
+  return state[NameSpace.DATA].offers[state[NameSpace.CITY].city];
 };
 
 export {getCities, getCityOffers};
